pyproject.toml
[tool.poetry]
name = "base-crud-fastapi"
version = "0.1.0"
description = ""
authors = ["vlad-terehoff <tereh_off95@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy-file = "^0.6.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.3.4"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
alembic-postgresql-enum = "^1.3.0"
fastapi-filters = "^0.2.9"
fastapi-pagination = "^0.12.26"
fasteners = "^0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
